rewrite importTiddlers function
The existing function relies on the html of a tiddlywiki being in
the correct format. Sometimes the html of a TiddlyWiki can be unusual, for
instance when the html of an iframe in IE6 is accessed the tags in the html
string are in upper case.

The previous method relies on doing a string search using locateStoreArea
which in the above situation will fail to locate the storeArea even though
the html is valid.

This method using jQuery is shorter and more reliable as it turns the
elements into dom elements and looks for the element with id storeArea which
is much more reliable.

This has been tested in IE6, Firefox and Chrome and all seem to work.