add TiddlyWebSaver.js as a library
TiddlyWebSaver.js is TiddlyWebConfig, ServerSideSavingPlugin,  TiddlyWebAdaptor
packaged into a single javascript file which makes use of tiddlywiki.ready to
delay loading until after the loading of plugins.

Note currently the TiddlyWebAdaptor works out where to save to from html in the
tiddlywiki itself. This requires a change in tiddlywebwiki. Previously this information
was read from the tiddler containing the adaptor plugin.