tweak TiddlyWebConfig and mechanism for getting host/workspace
there is no need for TiddlyWebSaver.jar
but since TiddlyWebConfig is no longer a plugin tiddler it needs to
get host and workspace from elsewhere.

For this we read directly from the dom in the template.
The template has an example host and workspace for demonstration
purposes. It is anticipated that a serverside would replace the contents
with its own config.